move schema/type def funcs to `./geo-type`
