+++ src/geo-type/map-tile.ts
@@ -1,5 +1,5 @@
-import { Type } from "./typebox.js";
-import type { SchemaOptions } from "./typebox.js";
+import { Type } from "../typebox.js";
+import type { SchemaOptions } from "../typebox.js";
 
 export const MIN_ZOOM = 0;
 export const MAX_ZOOM = 32;
